trim vue-query example
