--- examples/vue-query/pages/about/+Page.vue
@@ -1,4 +1,0 @@
-<template>
-  <h1>About</h1>
-  <p>Example of using Vike.</p>
-</template>
