--- examples/vue-query/pages/about/+config.ts
@@ -1,5 +1,0 @@
-import type { Config } from 'vike/types'
-
-export default {
-  title: 'About - My Vike + Vue App'
-} satisfies Config
