+++ examples/vue-query/pages/index/+title.ts
