+++ examples/vue-query/pages/index/types.ts
