--- examples/vue-query/pages/star-wars/movie/+route.ts
@@ -1,1 +1,0 @@
-export default '/star-wars/@movieId'
