--- examples/vue-query/pages/star-wars/movie/+title.ts
@@ -1,9 +1,0 @@
-export { title }
-
-import type { Data } from './+data'
-import type { PageContext } from 'vike/types'
-
-function title(pageContext: PageContext<Data>) {
-  const movie = pageContext.data
-  return movie.title
-}
